+++ src/api.ts
@@ -109,8 +109,8 @@ class BigQueryAPIClient implements BigQueryAPI {
       .replace(
         query.rawSql,
         // The $__interval vars are hardcoded here, as validation is not performed via PanelQueryRunner which provides the interval value.
-        // Moreover, the interval and the rest of the query options are not passed down to the QueryEditor are these are dependand on
-        // dashbaord settings, panel size etc. Hardcoding the interval value here should not influence the query validation whatsoever.
+        // Moreover, the interval and the rest of the query options are not passed down to the QueryEditor are these are dependant on
+        // dashboard settings, panel size etc. Hardcoding the interval value here should not influence the query validation whatsoever.
         {
           __interval: {
             text: '1m',